Merge outstanding home/work changes.
The google_it function now works in visual mode, and uses the get_visual_selection function to get the selected word or line. The grep_buffers function has been re-written to prevent each buffer from opening in a new split. The 'complete' setting was tweaked to improve performance and the 'list' setting was turned back on.